HOTFIX: 미션 성공 여부 뷰 위치 수정
