+++ mission/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from .views import MissionListView, MissionImageUploadView, RandomMissionCreateView, MissionCheckCompleteView
+from .views import MissionListView, MissionImageUploadView, RandomMissionCreateView, MissionCheckCompleteView, IsMissionCompleteView
 
 urlpatterns = [
     path('list/', MissionListView.as_view({
@@ -16,4 +16,7 @@ urlpatterns = [
     path('check_complete/', MissionCheckCompleteView.as_view({
         'post': 'create',  # 사진을 올리고 검사를 받는 로직 구성
     }), name='mission_check'),
+    path('is_complete/<int:pk>', IsMissionCompleteView.as_view({
+        'get': 'retrieve',
+    }))
 ]
