+++ mission/views.py
@@ -190,3 +190,19 @@ class RandomMissionCreateView(viewsets.ViewSet):
             "message": "랜덤 미션 할당 완료",
             "missions": created_missions
         }, status=status.HTTP_201_CREATED)
+
+
+class IsMissionCompleteView(viewsets.ViewSet):
+
+    def retrieve(self, request, *args, **kwargs):
+        tdp = kwargs.get('pk', None)
+        travel_days_and_places = None
+        try:
+            travel_days_and_places = TravelDaysAndPlaces.objects.get(id=tdp)
+        except TravelDaysAndPlaces.DoesNotExist:
+            return Response({'ERROR': '해당 여행 장소를 찾을 수 없습니다.'}, status=status.HTTP_404_NOT_FOUND)
+
+        return Response({
+            'tdp_id': tdp,
+            'mission_success': travel_days_and_places.mission_success
+        }, status=status.HTTP_200_OK)
