+++ tour/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from .views import TravelViewSet, NearEventView, AddTravelerView, GetAreaList, Sido_list, CourseView, IsMissionCompleteView
+from .views import TravelViewSet, NearEventView, AddTravelerView, GetAreaList, Sido_list, CourseView
 
 urlpatterns = [
     path('', TravelViewSet.as_view({
@@ -38,8 +38,5 @@ urlpatterns = [
         'get': 'retrieve',  # 개별 조회
         'delete': 'destroy' # 삭제
     }), name='course-detail'),
-    path('is_mission_complete/<int:pk>', IsMissionCompleteView.as_view({
-        'get': 'retrieve',
-    }))
 
 ]
